Connect_database_config/requirements.txt
mysql-connector-python
pyspark
kafka-python
python-dotenv
pymongo
redis
findspark